Changed some records to final classes to simplify the mediator implementation

+++ src/main/java/org/n3gd0r/recipe/domain/Mass.java
@@ -5,16 +5,16 @@ import org.springframework.util.Assert;
 /**
  * Mass
  */
-public record Mass(float value) {
+public record Mass(int value) {
     public Mass {
-        Assert.isTrue(value > 0.0, "The mass value shuold be a positive number");
+        Assert.isTrue(value > 0, "The mass value shuold be a positive number");
     }
 
-    public static Mass ofGrams(float value) {
+    public static Mass ofGrams(int value) {
         return new Mass(value);
     }
 
-    public static Mass ofKilograms(float value) {
+    public static Mass ofKilograms(int value) {
         return new Mass(value * 1000);
     }
 }
